+++ pyproject.toml
@@ -92,6 +92,9 @@ langsmith = [
 pdf = [
     "pymupdf>=1.25.0",
 ]
+olostep = [
+    "olostep>=0.1.0",
+]
 dev = [
     "pytest>=9.0.0,<10.0.0",
     "pytest-asyncio>=1.3.0,<2.0.0",
